Rename nodelink to MultiLink
