+++ README.md
@@ -1,18 +1,18 @@
-# Node-link View
+# MultiLink
 
-This is the node-link data visualization app for the Multinet project.
-It is a typescript application, written using the Vue.js framework, that allows
-users to visualize network data hosted on a Multinet instance.
+This is the MultiLink data visualization app for the Multinet project.
+It is a typescript node-link application, written using the Vue.js framework,
+that allows users to visualize network data hosted on a Multinet instance.
 
 ## Quick Start
 
 To get the application up and running in dev mode:
 
 1. Ensure that you have Node and yarn installed.
-2. Check out this repository and move into it (e.g., `cd view-nodelink`).
+2. Check out this repository and move into it (e.g., `cd MultiLink`).
 3. Copy .env.development to .env with `cp .env.development .env`.
 4. Install the Yarn dependencies: `yarn install`.
 5. Run the application: `yarn serve`.
 
-For further details, including how to set up a full multinet instance, please
+For further details, including how to set up a full Multinet instance, please
 see the [full documentation](https://multinet-app.readthedocs.io).
