+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "vue-nodelink",
+  "name": "MultiLink",
   "version": "0.1.0",
   "private": true,
-  "description": "A node-link visualization built to fit in with multinet",
+  "description": "",
   "author": "Visualization Design Lab <vdl.sci.utah.edu>",
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -58,5 +58,5 @@
     "> 1%",
     "last 2 versions"
   ],
-  "repository": "https://github.com/multinet-app/view-nodelink"
+  "repository": "https://github.com/multinet-app/MultiLink"
 }
