+++ src/components/MultiLink/Legend.vue
